+++ presentation/src/main/kotlin/com/lukelorusso/presentation/extensions/BitmapExtension.kt
@@ -47,7 +47,7 @@ fun createBitmap(width: Int, height: Int, color: Int): Bitmap {
     val bitmap = Bitmap.createBitmap(width, height, Bitmap.Config.ARGB_8888)
     val canvas = Canvas(bitmap)
     val paint = Paint()
-    paint.setColor(color)
+    paint.color = color
     canvas.drawRect(0f, 0f, width.toFloat(), height.toFloat(), paint)
     return bitmap
 }
