--- presentation/src/main/kotlin/com/lukelorusso/presentation/extensions/BooleanExtension.kt
@@ -1,7 +1,0 @@
-package com.lukelorusso.presentation.extensions
-
-/**
- * Zero is used to represent false, and One is used to represent true
- */
-fun Boolean.toInt(): Int =
-    if (this) 1 else 0
