+++ presentation/src/main/kotlin/com/lukelorusso/presentation/extensions/ViewExtension.kt
@@ -1,9 +1,7 @@
 package com.lukelorusso.presentation.extensions
 
-import android.os.Build
-import android.view.*
-import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.lukelorusso.presentation.R
+import android.view.View
+import android.view.ViewTreeObserver
 
 fun View.setAlphaWithAnimation(
     start: Float,
@@ -12,34 +10,6 @@ fun View.setAlphaWithAnimation(
     animationEnd: (() -> Unit)? = null
 ) = (start to end).applyFloatAnimation(duration = duration, animationEnd = animationEnd) {
     alpha = it
-}
-
-fun View.fadeInView(duration: Int? = null) {
-    if (this.visibility == View.VISIBLE) return
-    this.animation?.setAnimationListener(null)
-    this.animation?.cancel()
-    this.clearAnimation()
-    val effectDuration = duration ?: resources.getInteger(R.integer.fading_effect_duration_default)
-    this.visibility = View.INVISIBLE // setting precondition (just in case)
-    this.setAlphaWithAnimation(0F, 1F, effectDuration.toLong()) // setting animation
-    if (this is FloatingActionButton) this.show() // custom behaviour
-    else this.visibility = View.VISIBLE // making this visible to show the animation to the user
-}
-
-fun View.fadeOutView(duration: Int? = null) {
-    if (this.visibility == View.INVISIBLE) return
-    this.animation?.setAnimationListener(null)
-    this.animation?.cancel()
-    this.clearAnimation()
-    val effectDuration = duration ?: resources.getInteger(R.integer.fading_effect_duration_default)
-    this.visibility = View.VISIBLE // setting precondition (just in case)
-    this.setAlphaWithAnimation(
-        1F,
-        0F,
-        effectDuration.toLong()
-    ) // setting animation with listener at the end (if reached)
-    if (this is FloatingActionButton) this.hide() // custom behaviour
-    else this.visibility = View.INVISIBLE // making this invisible
 }
 
 fun View.addOneTimeOnGlobalLayoutListener(listener: () -> Unit) {
@@ -51,16 +21,4 @@ fun View.addOneTimeOnGlobalLayoutListener(listener: () -> Unit) {
     this.viewTreeObserver.addOnGlobalLayoutListener(
         onGlobalLayoutListener
     )
-}
-
-@Suppress("DEPRECATION")
-fun View.statusBarHeight() = when {
-    Build.VERSION.SDK_INT >= Build.VERSION_CODES.R ->
-        rootWindowInsets.getInsetsIgnoringVisibility(WindowInsets.Type.systemBars()).top
-
-    Build.VERSION.SDK_INT >= Build.VERSION_CODES.M ->
-        rootWindowInsets?.systemWindowInsetTop ?: 0
-
-    else ->
-        0
 }
