+++ presentation/src/main/kotlin/com/lukelorusso/presentation/ui/capture/CaptureUiState.kt
@@ -1,9 +1,9 @@
-package com.lukelorusso.presentation.ui.camera
+package com.lukelorusso.presentation.ui.capture
 
 import com.lukelorusso.domain.model.Color
 import com.lukelorusso.presentation.ui.base.ContentState
 
-data class CameraUiState(
+data class CaptureUiState(
     val contentState: ContentState = ContentState.CONTENT,
     val lastLensPosition: Int? = null,
     val lastZoomValue: Int? = null,
