--- presentation/src/main/res/anim/transition_enter_from_left.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<set xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shareInterpolator="false">
-    <translate
-        android:duration="@integer/fading_effect_duration_default"
-        android:fromXDelta="100%"
-        android:fromYDelta="0%"
-        android:toXDelta="0%"
-        android:toYDelta="0%" />
-</set>
