--- presentation/src/main/res/drawable/drag_handle_white.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="25dp"
-    android:height="25dp"
-    android:tint="@color/white_FFFFFF"
-    android:viewportWidth="24.0"
-    android:viewportHeight="24.0">
-    <path
-        android:fillColor="#FF000000"
-        android:pathData="M20,9H4v2h16V9zM4,15h16v-2H4v2z" />
-</vector>
