--- presentation/src/main/res/drawable/mid_placeholder.xml
@@ -1,18 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item>
-        <shape android:shape="rectangle">
-            <size
-                android:width="30dp"
-                android:height="30dp" />
-        </shape>
-    </item>
-
-    <item
-        android:bottom="5dp"
-        android:drawable="@drawable/drag_handle_white"
-        android:left="5dp"
-        android:right="5dp"
-        android:top="5dp" />
-
-</layer-list>
