--- presentation/src/main/res/drawable/min_placeholder.xml
@@ -1,21 +1,0 @@
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item>
-        <shape android:shape="oval">
-            <size
-                android:width="27dp"
-                android:height="27dp" />
-            <stroke
-                android:width="2dp"
-                android:color="@color/white_FFFFFF" />
-        </shape>
-    </item>
-
-    <item
-        android:bottom="5dp"
-        android:drawable="@drawable/zoom_minus_white"
-        android:left="5dp"
-        android:right="5dp"
-        android:top="5dp" />
-
-</layer-list>
