--- presentation/src/main/res/drawable/shape_bottom_toolbar.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/black_50" />
-
-    <corners
-        android:bottomLeftRadius="0dp"
-        android:bottomRightRadius="0dp"
-        android:topLeftRadius="@dimen/bottom_sheet_corner_radius"
-        android:topRightRadius="@dimen/bottom_sheet_corner_radius" />
-
-</shape>
