--- presentation/src/main/res/layout/fragment_camera.xml
@@ -1,79 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.camera.CameraFragment">
-
-    <FrameLayout
-        android:id="@+id/content"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content" />
-
-    <io.fotoapparat.view.CameraView
-        android:id="@+id/cameraView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <!--Optionally add tap to focus-->
-        <io.fotoapparat.view.FocusView
-            android:id="@+id/focusView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </io.fotoapparat.view.CameraView>
-
-    <ImageView
-        android:id="@+id/cameraViewfinder"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:contentDescription="@null"
-        app:srcCompat="@drawable/viewfinder" />
-
-    <include
-        android:id="@+id/inclToolbarCameraTop"
-        layout="@layout/layout_camera_top_toolbar"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-    <include
-        android:id="@+id/inclToolbarCameraBottom"
-        layout="@layout/layout_camera_bottom_toolbar" />
-
-    <com.lukelorusso.verticalseekbar.VerticalSeekBar
-        android:id="@+id/cameraZoomSeekBar"
-        android:layout_width="70dp"
-        android:layout_height="match_parent"
-        android:layout_above="@id/inclToolbarColor"
-        android:layout_below="@id/inclToolbarCameraTop"
-        android:layout_alignParentEnd="true"
-        android:layout_centerVertical="true"
-        android:layout_gravity="center"
-        android:layout_marginTop="15dp"
-        android:layout_marginBottom="15dp"
-        android:visibility="invisible"
-        app:vsb_bar_background="@color/white_FFFFFF"
-        app:vsb_bar_progress_gradient_end="@color/color_primary"
-        app:vsb_bar_progress_gradient_start="@color/color_accent"
-        app:vsb_bar_width="2dp"
-        app:vsb_click_to_set_progress="true"
-        app:vsb_max_placeholder_position="outside"
-        app:vsb_max_placeholder_src="@drawable/max_placeholder"
-        app:vsb_min_placeholder_position="outside"
-        app:vsb_min_placeholder_src="@drawable/min_placeholder"
-        app:vsb_show_thumb="true"
-        app:vsb_thumb_container_corner_radius="10dp"
-        app:vsb_thumb_container_tint="@color/color_primary"
-        app:vsb_thumb_placeholder_src="@drawable/mid_placeholder"
-        app:vsb_use_thumb_to_set_progress="true" />
-
-    <include
-        android:id="@+id/inclToolbarColor"
-        layout="@layout/layout_color_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@id/inclToolbarCameraBottom" />
-
-</RelativeLayout>
