--- presentation/src/main/res/layout/layout_camera_bottom_toolbar.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="85dp"
-    android:layout_alignParentBottom="true"
-    android:background="@color/black_50"
-    android:clickable="true"
-    android:focusable="true"
-    android:focusableInTouchMode="true"
-    android:paddingLeft="10dp"
-    android:paddingRight="10dp">
-
-    <ImageButton
-        android:id="@+id/toolbarInfoButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_centerVertical="true"
-        android:background="@color/transparent"
-        android:padding="10dp"
-        app:srcCompat="@drawable/info_outline_white" />
-
-    <ImageButton
-        android:id="@+id/toolbarCameraButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:background="@color/transparent"
-        app:srcCompat="@drawable/camera_big_white" />
-
-    <include
-        android:id="@+id/toolbarProgressBar"
-        layout="@layout/layout_progress"
-        android:layout_width="@dimen/progress_bar_dimens"
-        android:layout_height="@dimen/progress_bar_dimens"
-        android:layout_centerInParent="true"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-    <ImageButton
-        android:id="@+id/toolbarHistoryButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_centerVertical="true"
-        android:background="@color/transparent"
-        android:padding="10dp"
-        app:srcCompat="@drawable/history_big_white" />
-
-</RelativeLayout>
