--- presentation/src/main/res/layout/layout_camera_top_toolbar.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@color/black_50"
-    android:orientation="vertical"
-    android:clickable="true"
-    android:focusable="true"
-    android:focusableInTouchMode="true"
-    android:paddingLeft="10dp"
-    android:paddingRight="10dp">
-
-    <View
-        android:id="@+id/toolbarTopSpacer"
-        android:layout_width="match_parent"
-        android:layout_height="26dp" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:paddingTop="10dp"
-        android:paddingBottom="10dp"
-        android:layout_gravity="center_vertical">
-
-        <ImageButton
-            android:id="@+id/toolbarSwitchCameraButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@color/transparent"
-            android:padding="10dp"
-            android:visibility="gone"
-            app:srcCompat="@drawable/camera_front_white"
-            tools:visibility="visible" />
-
-        <View
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_weight="1" />
-
-        <ImageButton
-            android:id="@+id/toolbarFlashButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@color/transparent"
-            android:padding="10dp"
-            android:visibility="gone"
-            app:srcCompat="@drawable/flash_on_white"
-            tools:visibility="visible" />
-
-    </LinearLayout>
-
-</LinearLayout>
