--- presentation/src/main/res/layout/layout_color_toolbar.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="@drawable/shape_bottom_toolbar"
-    android:padding="10dp"
-    android:visibility="invisible"
-    tools:visibility="visible">
-
-    <View
-        android:id="@+id/colorPreviewPanel"
-        android:layout_width="@dimen/color_picker_dimens"
-        android:layout_height="@dimen/color_picker_dimens"
-        android:layout_alignParentEnd="true"
-        android:layout_centerVertical="true"
-        android:background="@drawable/circle_color_picker" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/colorMainLine"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white_FFFFFF"
-            android:textSize="18sp"
-            android:textStyle="bold"
-            tools:text="Marrone Bistro" />
-
-        <TextView
-            android:id="@+id/colorTopLine"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white_FFFFFF"
-            android:textSize="14sp"
-            tools:text="#RRGGBB" />
-
-        <TextView
-            android:id="@+id/colorBottomLine"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/white_FFFFFF"
-            android:textSize="14sp"
-            tools:text="rgb(rr.rr%, gg.gg%, bb.bb%)" />
-
-    </LinearLayout>
-
-</RelativeLayout>
